+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - 2021-01-08
 
+### Added
+
+- Exclude flag to ignore specific directories
+
 ### Changed
 
 - Switched to Cobra CLI